# docs/openapi.yaml
openapi: 3.0.3
info:
  title: API Documentation
  description: Документация API для всех операций, включая аутентификацию, управление картами, блокировку карт и управление пользователями
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Локальный сервер для разработки
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    BaseUserDto:
      type: object
      description: Базовый объект запроса для данных пользователя
      required:
        - login
        - password
      properties:
        login:
          type: string
          description: Логин пользователя
          example: user
          minLength: 3
          maxLength: 20
          pattern: ^[a-zA-Z0-9_]+$
        password:
          type: string
          description: Пароль пользователя
          example: Passw0rd*
          minLength: 8
          pattern: ^(?=.*[0-9])(?=.*[a-z])(?=.*[A-Z])(?=.*[!@#&()–[{}]:;',?/*~$^+=<>]).*$
    AuthRequestDto:
      allOf:
        - $ref: '#/components/schemas/BaseUserDto'
      description: Запрос для авторизации пользователя
    AuthResponseDto:
      type: object
      description: Объект ответа аутентификации, содержащий JWT-токен и информацию о пользователе
      required:
        - token
        - user
      properties:
        token:
          type: string
          description: JWT-токен для аутентификации
          example: eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJ2aWxhcnJlYWxpdHkiLCJpYXQiOjE3NTQ5ODg5MjYsImV4cCI6MTc1NTA3NTMyNn0.DnEr5Gx5xuaD0Kf0x_SFoYm6QGazn61UNNfJHCqkRB3
        user:
          $ref: '#/components/schemas/UserResponseDto'
    UserResponseDto:
      type: object
      description: Объект ответа, содержащий информацию о пользователе
      required:
        - id
        - login
        - roleName
        - isEnabled
        - isNonLocked
      properties:
        id:
          type: integer
          format: int64
          description: Уникальный идентификатор пользователя
          example: 10
        login:
          type: string
          description: Логин пользователя
          example: user123
        roleName:
          type: string
          description: Название роли пользователя
          example: ADMIN
        isEnabled:
          type: boolean
          description: Флаг, указывающий, активен ли пользователь
          example: true
        isNonLocked:
          type: boolean
          description: Флаг, указывающий, не заблокирован ли пользователь
          example: true
        cards:
          type: array
          description: Список карт пользователя (может отсутствовать в некоторых ответах)
          items:
            $ref: '#/components/schemas/CardResponseDto'
...
    CardResponseDto:
      type: object
      description: Объект ответа, содержащий информацию о карте
      required:
        - id
        - maskedCardNumber
        - ownerLogin
        - expiryDate
        - statusName
        - balance
      properties:
        id:
          type: integer
          format: int64
          description: Уникальный идентификатор карты
          example: 10
        maskedCardNumber:
          type: string
          description: Маскированный номер карты, отображающий только последние 4 цифры
          example: "************1234"
        ownerLogin:
          type: string
          description: Логин владельца карты
          example: user123
        expiryDate:
          type: string
          pattern: ^\d{4}-\d{2}$
          description: Дата истечения срока действия карты в формате yyyy-MM
          example: 2025-08
        statusName:
          type: string
          description: Название статуса карты
          example: ACTIVE
        balance:
          type: number
          format: double
          description: Текущий баланс на карте
          example: 1500.75
...
paths:
  /register:
    post:
      summary: Регистрация нового пользователя
      description: Создает нового пользователя и возвращает JWT-токен и информацию о пользователе
      tags:
        - Аутентификация
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRequestDto'
      responses:
        '200':
          description: Успешная регистрация
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponseDto'
        '400':
          description: Некорректные данные запроса или ошибка разбора токена
          content: {}
        '401':
          description: Ошибка аутентификации
          content: {}
        '403':
          description: Доступ запрещен
          content: {}
        '409':
          description: Пользователь с таким логином уже существует
          content: {}
        '500':
          description: Внутренняя ошибка сервера
          content: {}
  /login:
    post:
      summary: Вход пользователя
      description: Аутентифицирует пользователя и возвращает JWT-токен и информацию о пользователе
      tags:
        - Аутентификация
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRequestDto'
      responses:
        '200':
          description: Успешный вход
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponseDto'
        '400':
          description: Некорректные данные запроса или ошибка разбора токена
          content: {}
        '401':
          description: Неверный логин или пароль / ошибка аутентификации
          content: {}
        '403':
          description: Доступ запрещен
          content: {}
        '500':
          description: Внутренняя ошибка сервера
          content: {}
  /me:
    get:
      summary: Получение информации о текущем пользователе
      description: Возвращает информацию о текущем пользователе на основе JWT-токена
      tags:
        - Аутентификация
      security:
        - BearerAuth: []
      parameters:
        - name: Authorization
          in: header
          required: true
          schema:
            type: string
            example: Bearer eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJ2aWxhcnJlYWxpdHkiLCJpYXQiOjE3NTQ5ODg5MjYsImV4cCI6MTc1NTA3NTMyNn0.DnEr5Gx5xuaD0Kf0x_SFoYm6QGazn61UNNfJHCqkRB3
          description: JWT-токен в формате "Bearer <token>"
      responses:
        '200':
          description: Успешное получение информации о пользователе
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponseDto'
        '400':
          description: Ошибка разбора токена
          content: {}
        '401':
          description: Недействительный или отсутствующий токен / ошибка аутентификации
          content: {}
        '403':
          description: Доступ запрещен
          content: {}
        '500':
          description: Внутренняя ошибка сервера
          content: {}
...